fix tables

+++ internal/app/domain/sensor.go
@@ -8,7 +8,7 @@ import (
 
 type Sensor struct {
 	ID            uint           `gorm:"primaryKey" json:"id"`
-	Esp32DeviceID uint64         `gorm:"not null" json:"esp32_device_id"`
+	Esp32DeviceID uint           `gorm:"not null" json:"esp32_device_id"`
 	Esp32Device   Esp32Device    `gorm:"foreignKey:Esp32DeviceID" json:"esp32_device"`
 	ParkingLotID  uint           `gorm:"not null" json:"parking_lot_id"`
 	ParkingLot    ParkingLot     `gorm:"foreignKey:ParkingLotID" json:"parking_lot"`
